improved scrolling behavior of pipeline details page - right half now scrolls independently from the left half
