+++ apps/generic-issuance-client/src/components/Core.tsx
@@ -46,5 +46,8 @@ export const HomeLink = (): JSX.Element => {
 };
 
 export const PageContent = styled.div`
+  margin-top: 64px;
   padding: 32px;
+  width: 100%;
+  height: 100%;
 `;
