+++ apps/generic-issuance-client/src/pages/pipeline/DetailsSections/PipelineAddManualTicketSection.tsx
@@ -167,6 +167,7 @@ ticket: ${ticketTypeName}`
       <FormControl mb={2}>
         <FormLabel>Event</FormLabel>
         <Select
+          maxW={"100%"}
           w="sm"
           mt={2}
           value={eventId}
@@ -183,6 +184,7 @@ ticket: ${ticketTypeName}`
       <FormControl mb={2}>
         <FormLabel>Ticket Type</FormLabel>
         <Select
+          maxW={"100%"}
           w="sm"
           mt={2}
           value={ticketTypeId}
@@ -204,6 +206,7 @@ ticket: ${ticketTypeName}`
           placeholder="Bob Glob"
           type="text"
           width="sm"
+          maxW={"100%"}
         />
       </FormControl>
 
@@ -216,6 +219,7 @@ ticket: ${ticketTypeName}`
           placeholder="email@provider.tld"
           type="email"
           width="sm"
+          maxW={"100%"}
           autoComplete="off"
           data-1p-ignore
         />
@@ -225,6 +229,7 @@ ticket: ${ticketTypeName}`
         colorScheme="blue"
         mt={1}
         width="sm"
+        maxW={"100%"}
         onClick={onAddClick}
         isLoading={inProgress}
       >
