+++ apps/generic-issuance-client/src/pages/pipeline/SectionContainer.tsx
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+export const SectionContainer = styled.div`
+  border-radius: 8px;
+  background-color: rgba(255, 255, 255, 0.03);
+  border: 1px solid #4b4b4b;
+  min-width: 8px;
+  padding: 16px;
+  display: block;
+  overflow-x: hidden;
+`;
